requirements.txt
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.3.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pygame==2.5.2
pypiwin32==223
pyttsx3==2.90
pywin32==306
requests==2.31.0
rsa==4.9
srt==3.5.3
tqdm==4.66.2
urllib3==2.2.1
vosk==0.3.45
websockets==12.0
